Fixed REST API tests (WIP)
