+++ src/test/scala/com/ltonetwork/http/AddressRouteSpec.scala
@@ -46,24 +46,6 @@ class AddressRouteSpec
     message <- Gen.listOfN(length, Gen.alphaNumChar).map(_.mkString).label("message")
   } yield (account, message)
 
-  routePath("/seq/{from}/{to}") in {
-    val r1 = Get(routePath("/seq/1/4")) ~> route ~> check {
-      val response = responseAs[Seq[String]]
-      response.length shouldBe 3
-      allAddresses should contain allElementsOf response
-      response
-    }
-
-    val r2 = Get(routePath("/seq/5/9")) ~> route ~> check {
-      val response = responseAs[Seq[String]]
-      response.length shouldBe 4
-      allAddresses should contain allElementsOf response
-      response
-    }
-
-    r1 shouldNot contain allElementsOf r2
-  }
-
   routePath("/validate/{address}") in {
     val t = Table(("address", "valid"), allAddresses.map(_ -> true) :+ "invalid-address" -> false: _*)
 
@@ -75,36 +57,6 @@ class AddressRouteSpec
       }
     }
   }
-
-  routePath("/seed/{address}") in {
-    val account = allAccounts.head
-    val path    = routePath(s"/seed/${account.address}")
-    Get(path) ~> route should produce(ApiKeyNotValid)
-    Get(path) ~> api_key(apiKey) ~> route ~> check {
-      val json = responseAs[JsObject]
-      (json \ "address").as[String] shouldEqual account.address
-      (json \ "seed").as[String] shouldEqual Base58.encode(account.seed)
-    }
-  }
-
-  private def testSign(path: String, encode: Boolean): Unit =
-    forAll(generatedMessages) {
-      case (account, message) =>
-        val uri = routePath(s"/$path/${account.address}")
-        Post(uri, message) ~> route should produce(ApiKeyNotValid)
-        Post(uri, message) ~> api_key(apiKey) ~> route ~> check {
-          val resp      = responseAs[JsObject]
-          val signature = Base58.decode((resp \ "signature").as[String]).get
-
-          (resp \ "message").as[String] shouldEqual (if (encode) Base58.encode(message.getBytes) else message)
-          (resp \ "publicKey").as[String] shouldEqual Base58.encode(account.publicKey)
-
-          crypto.verify(signature, message.getBytes, account.publicKey) shouldBe true
-        }
-    }
-
-  routePath("/sign/{address}") in testSign("sign", true)
-  routePath("/signText/{address}") in testSign("signText", false)
 
   private def testVerify(path: String, encode: Boolean): Unit = {
 
@@ -134,19 +86,6 @@ class AddressRouteSpec
 
   routePath("/verifyText/{address}") in testVerify("verifyText", false)
   routePath("/verify/{address}") in testVerify("verify", true)
-
-  routePath("") in {
-    Post(routePath("")) ~> route should produce(ApiKeyNotValid)
-    Post(routePath("")) ~> api_key(apiKey) ~> route ~> check {
-      allAddresses should not contain (responseAs[JsObject] \ "address").as[String]
-    }
-  }
-
-  routePath("/{address}") in {
-    Delete(routePath(s"/${allAddresses.head}")) ~> api_key(apiKey) ~> route ~> check {
-      (responseAs[JsObject] \ "deleted").as[Boolean] shouldBe true
-    }
-  }
 
   routePath(s"/scriptInfo/${allAddresses(1)}") in {
     (blockchain.accountScript _).when(allAccounts(1).toAddress).onCall((_: Address) => Some(ScriptV1(TRUE).explicitGet()))
